Update script.lua
Completely remade the way I check data (at the recommendation of fade), leading myself to create the checkCyclic function which iterates through all data sent from the remote/bindable, strips all metatables, checks if any of the arguments are cyclic tables, and proceeds to not process them if they are.  This function also checks to ensure it wont overflow the stack, and if it is going to, it passes the error to roblox, giving the exact intended outcome.