Update scriptBeta.lua
fixed an issue where task.defer would error when over 8k args were passed.  Also added an arg limiter so the remotespy won't render in an infinite amount of arguments, and I made it so the remotespy won't log any calls from callbacks if they are not called by a cclosure (preventing a game from setting the callback to a function, and then calling that function with impossible args).