Update scriptBeta.lua
Changed the way garbage collection of the UI works, making it purely rely on Synapse's garbage collection, which causes the Remote Spy to take anywhere from 5~30 seconds to clean up and close.  Also added in Get Call Stack (allowing the user to see the stack of calls leading to the remote call), and Get Calling Script V2, which returns the most recent script in the Call Stack.  These are both extremely useful for games that use module scripts, where getcallingscript (the Synapse function) doesn't return module scripts, and where the origination of call arguments can be confusing.  This update also changed how ShallowClone works, getting rid of useless arguments earlier on, and preventing some issues that would occur later when a function was passed as an argument to a FireServer call.