Update scriptBeta.lua
rolled back old change because of stuff being stupid.  I tested out oldIndex(remote, "ClassName") in a pcall, and it broke the pcall, making it not return at all, but it would error "Unable to assign property Value. Object expected, got String."  An error which makes no sense to me given the context.